Update README.md
 California Code of Regulations

Welcome to the newly enhanced site for the California Code of Regulations. This site has been upgraded to assure you a positive Thomson Reuters Westlaw experience. This website is maintained by Thomson Reuters.

By using this website, you agree not to use it in any manner that could disable, overburden, damage, or impair the site or interfere with any other party’s use of the website, or to use any device, software or routine that interferes with the proper working of the website.